+++ packages/launcher/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @addfox/launcher
 
+## 0.2.3
+
+### Patch Changes
+
+- chore: resolve latest addfox versions from npm registry to avoid pnpm lowest-direct resolution
+
 ## 0.2.2
 
 ### Patch Changes